Migrate #testPutConnectorConfigProducerError to Mockito
